requirements.txt
pandas 
numpy
matplotlib
seaborn
sentence-transformers
transformers
faiss-cpu
torch
google-generativeai
python-dotenv
tqdm
streamlit

